External partition: Use int64_t for tpl features

+++ aom/aom_external_partition.h
@@ -165,9 +165,9 @@ typedef enum {
 typedef struct aom_sb_tpl_features {
   int tpl_unit_length; /**< The block length of tpl process */
   int num_units;       /**< The number of units inside the current superblock */
-  double intra_cost[64];  /**< The intra cost of each unit */
-  double inter_cost[64];  /**< The inter cost of each unit */
-  double mc_dep_cost[64]; /**< The motion compensated dependecy cost */
+  int64_t intra_cost[64];  /**< The intra cost of each unit */
+  int64_t inter_cost[64];  /**< The inter cost of each unit */
+  int64_t mc_dep_cost[64]; /**< The motion compensated dependecy cost */
 } aom_sb_tpl_features_t;
 
 /*!\brief Features collected from the simple motion process.
